BLonD v1.6.0. See README.md for details.

+++ beams/slices.py
@@ -73,7 +73,7 @@ class Slices(object):
             raise RuntimeError('The slicing_coord is not recognized')
         
         #: *Number of macroparticles per slice (~profile).*
-        self.n_macroparticles = np.zeros(n_slices)
+        self.n_macroparticles = np.zeros(n_slices, dtype='uint32')
         
         #: *Edges positions of the slicing*
         self.edges = np.zeros(n_slices + 1)
